Move all constants to the common.
