--- client/src/main/java/org/opensearch/ml/client/MLConstants.java
@@ -1,21 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package org.opensearch.ml.client;
-
-public class MLConstants {
-    // train, predict, trainandpredict
-    public static final String ACTION = "action";
-    public static final String TRAIN = "train";
-    public static final String PREDICT = "predict";
-    public static final String TRAINANDPREDICT = "trainandpredict";
-    // algorithm
-    public static final String ALGORITHM = "algorithm";
-    public static final String KMEANS = "kmeans";
-    public static final String RCF = "rcf";
-    // common parameters
-    public static final String ASYNC = "async";
-    public static final String MODELID = "modelid";
-}
